+++ src/presenters/includes/profile.js
@@ -62,23 +62,7 @@ InfoContainer.propTypes = {
   children: PropTypes.node.isRequired,
 };
 
-// Cover Container
 
-export const CoverContainer = ({ buttons, children, className, ...props }) => (
-  <div className={`cover-container ${className}`} {...props}>
-    {children}
-    {buttons}
-  </div>
-);
-CoverContainer.propTypes = {
-  buttons: PropTypes.node,
-  children: PropTypes.node.isRequired,
-  className: PropTypes.string,
-};
-CoverContainer.defaultProps = {
-  className: '',
-  buttons: null,
-};
 
 // Profile Container
